# Speed up convert to StandardModel by using threads