Converting models to `StandardModel` is pretty slow. I added a small improvement to use threads to process genes, reactions and metabolites on separate threads to speed this up. Slight improvements noted.

Results using iML1515 (2712 reactions, 1877 metabolites) as an example on my laptop using 4 threads (before, after timings):

iML1515.json: 7.2 seconds -> 5.4 seconds ~ 25% improvement
iML1515.mat: 1.4 seconds -> 1 second ~ 28% improvement
iML1515.xml: 5.6 seconds -> 5.5 seconds ~ 2% improvement

It seems to help for all the model types except SBML... No idea why it is different there...?