@exaexa 2021-05-21
For optimizations, we first need some detailed benchmark of where&how the performance is getting lost. I guess the most of the slowness comes from the (ugly) way of picking out the reaction IDs from json models, which can be fixed much more effectively than just with throwing additional cores at it. I'll get to it asap.