@stelmo 2021-05-21
> would this implementation also still work when I'd have only 1 thread?

Yes, all the tests are relative to the one thread case, which is the same speed as the original code. However, let's talk about threads more with @exaexa before going further, I don't think he was super supportive of adding this kind of functionality it.