@stelmo 2021-05-25
> I am not sure that it is a good idea to have the `@timeit` macro calls inside the `convert` function by default

Yep, I will remove these if we decide on going forward with this PR. @exaexa should we even bother with threads? If not, then we can delete this PR