@stelmo 2021-05-25
> Threads are better avoided here. Instead we should make the conversion efficient with the indexes. I'll make it work tomorrow.

Ah yeah, I promised to do that! Will start now!